requirements.txt
flask
flasgger
confluent_kafka
psycopg[binary]
python-keycloak